Add first files
This now functions as a Makefile clone with one target. It's not quite
at v0.1.0 because I'd like to make it a little cleaner, but it's very
close.

A buildfile is a ".cement" file that lists the commands to execute,
separated by a newline. Running `cement build` looks in the buildfile
and executes each command, separated by a newline and prints the output.
If a command fails, the rest are aborted and the process ends.

Once the process is finished, either by an earlier fail or by completing
all the tasks, the script prints "Build succeeded" or "Build failed" and
exits.